Added job and message microservice,added communication between microservices,and added methods in user and post microservices

+++ BackEnd/PostService/Controllers/BaseController.cs
@@ -27,7 +27,7 @@ namespace PostService.Controllers
         }
 
         [HttpGet("{id}")]
-        public IActionResult Get(int id)
+        public IActionResult Get(long id)
         {
             TEntity entity = _base.Get(id);
             if (entity == null)
